♻️ Refactor: store, review tsoa 적용 (#7)
