+++ src/index.ts
@@ -8,8 +8,8 @@ import { RegisterRoutes } from "./generated/routes.js";
 import { AppError } from "./common/errors/app.error.js";
 // import { handleUserSignUp } from "./modules/users/controllers/user.controller.js";
 // import { handleAddStore } from "./modules/stores/controllers/store.controller.js";
-import { handleAddReview, handleListMyReviews, handleListStoreReviews } from "./modules/reviews/controllers/review.controller.js";
-import { handleAddMission, handleListStoreMissions } from "./modules/missions/controllers/mission.controller.js";
+// import { handleListMyReviews } from "./modules/reviews/controllers/review.controller.js";
+import { handleAddMission } from "./modules/missions/controllers/mission.controller.js";
 import { handleChallengeMission, handleCompleteMission, handleListChallenge } from "./modules/missions/controllers/challenge.controller.js";
 
 // 1. 환경 변수 설정
@@ -49,12 +49,12 @@ app.get("/", (req: Request, res: Response) => {
 
 // app.post("/api/v1/users/signup", handleUserSignUp); // 회원가입
 // app.post("/api/v1/stores", handleAddStore); // 특정 지역에 가게 추가하기
-app.post("/api/v1/reviews", handleAddReview); // 가게에 리뷰 추가하기
+// app.post("/api/v1/reviews", handleAddReview); // 가게에 리뷰 추가하기
 app.post("/api/v1/missions", handleAddMission); // 가게에 미션 추가하기
 app.post("/api/v1/missions/challenge", handleChallengeMission); // 가게의 미션을 도전 중인 미션에 추가(미션 도전하기)
-app.get("/api/v1/stores/:storeId/reviews", handleListStoreReviews); // 가게 리뷰 조회하기
-app.get("/api/v1/reviews/:memberId", handleListMyReviews); // 내가 작성한 리뷰 조회하기
-app.get("/api/v1/stores/:storeId/missions", handleListStoreMissions); // 특정 가게의 미션 목록 조회
+// app.get("/api/v1/stores/:storeId/reviews", handleListStoreReviews); // 가게 리뷰 조회하기
+// app.get("/api/v1/reviews/:memberId", handleListMyReviews); // 내가 작성한 리뷰 조회하기
+// app.get("/api/v1/stores/:storeId/missions", handleListStoreMissions); // 특정 가게의 미션 목록 조회
 app.get("/api/v1/members/:memberId/missions", handleListChallenge); // 내가 진행 중인 미션 목록 조회
 app.patch("/api/v1/members/:missionId", handleCompleteMission); // 내가 진행 중인 미션을 진행 완료로 바꾸기
