+++ src/modules/missions/repositories/mission.repository.ts
@@ -20,27 +20,3 @@ export const getMission = async (missionId: number) => {
 
   return mission;
 };
-
-// 특정 가게의 미션 목록 조회
-export const getStoreMissions = async (store_id: number, cursor: number) => {
-  const missions = await prisma.mission.findMany({
-    select: {
-      id: true,
-      reward: true,
-      deadline: true,
-      mission_spec: true,
-    },
-    where: {
-      store_id,
-      id: {
-        gt: cursor,
-      },
-    },
-    orderBy: {
-      id: "asc",
-    },
-    take: 5,
-  });
-
-  return missions;
-};
