+++ src/modules/missions/services/mission.service.ts
@@ -1,5 +1,5 @@
-import { IAddMissionRequest, missionResponse, responseMissions } from "../dtos/mission.dto.js";
-import { addMission, getMission, getStoreMissions } from "../repositories/mission.repository.js";
+import { IAddMissionRequest, missionResponse } from "../dtos/mission.dto.js";
+import { addMission, getMission } from "../repositories/mission.repository.js";
 
 // 가게에 미션 추가하기
 export const addStoreMission = async (data: IAddMissionRequest) => {
@@ -21,19 +21,4 @@ export const addStoreMission = async (data: IAddMissionRequest) => {
     deadline: mission.deadline ? mission.deadline.toISOString() : undefined,
     missionSpec: mission.mission_spec || "",
   });
-};
-
-// 특정 가게의 미션 목록 조회
-export const handleStoreMissions = async (storeId: number, cursor: number) => {
-  const missions = await getStoreMissions(storeId, cursor);
-
-  const mapped = missions.map((m) => ({
-    id: Number(m.id),
-    storeId: storeId,
-    reward: Number(m.reward),
-    deadline: m.deadline?.toISOString(),
-    missionSpec: m.mission_spec || "",
-  }));
-
-  return responseMissions(mapped);
 };
