+++ src/modules/reviews/controllers/review.controller.ts
@@ -1,42 +1,25 @@
-import { Request, Response, NextFunction } from "express";
-import { StatusCodes } from "http-status-codes";
-import { addReviewRequest, IAddReviewRequest } from "../dtos/review.dto.js";
-import { addStoreReview, handleMyReviews, listStoreReviews } from "../services/review.service.js";
+import { IAddReviewRequest, IReviewResponse } from "../dtos/review.dto.js";
+import { addStoreReview, handleMyReviews } from "../services/review.service.js";
+import { Body, Controller, Get, Path, Post, Query, Route, Tags } from "tsoa";
+import { ApiResponse, success } from "../../../common/responses/response.js";
+import { ReviewListResponse } from "../../stores/dtos/store.dto.js";
 
 // 리뷰 작성하기
-export const handleAddReview = async (req: Request, res: Response, next: NextFunction) => {
-  console.log("리뷰 작성을 요청했습니다.");
-  console.log("body: ", req.body);
-
-  const review = await addStoreReview(addReviewRequest(req.body as IAddReviewRequest));
-
-  res.status(StatusCodes.OK).json({ result: review });
-};
-
-// 가게 리뷰 목록 조회
-export const handleListStoreReviews = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-  try {
-    const storeId = parseInt(req.params.storeId as string, 10);
-    const cursor = typeof req.query.cursor === "string" ? parseInt(req.query.cursor, 10) : 0;
-
-    const reviews = await listStoreReviews(storeId, cursor);
-
-    res.status(StatusCodes.OK).json(reviews);
-  } catch (err) {
-    next(err);
+@Route("reviews")
+@Tags("Reviews")
+export class ReviewController extends Controller {
+  @Post("")
+  public async handleAddReview(@Body() body: IAddReviewRequest): Promise<ApiResponse<IReviewResponse>> {
+    console.log("리뷰 작성을 요청했습니다.");
+    console.log("body: ", body);
+    const review = await addStoreReview(body);
+    return success(review);
   }
-};
 
-// 내가 작성한 리뷰 목록 조회하기
-export const handleListMyReviews = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-  try {
-    const memberId = parseInt(req.params.memberId as string, 10);
-    const cursor = typeof req.query.cursor === "string" ? parseInt(req.query.cursor, 10) : 0;
-
-    const reviews = await handleMyReviews(memberId, cursor);
-
-    res.status(StatusCodes.OK).json(reviews);
-  } catch (err) {
-    next(err);
+  // 내가 작성한 리뷰 목록 조회하기
+  @Get("{memberId}")
+  public async handleListMyReviews(@Path() memberId: number, @Query() cursor: number = 0): Promise<ApiResponse<ReviewListResponse>> {
+    const myReview = await handleMyReviews(memberId, cursor);
+    return success(myReview);
   }
-};
+}
