+++ src/modules/reviews/repositories/review.repository.ts
@@ -29,41 +29,6 @@ export const getReview = async (reviewId: number) => {
   return review;
 };
 
-// 가게 리뷰 목록 조회
-export const getAllStoreReviews = async (store_id: number, cursor: number) => {
-  const reviews = await prisma.review.findMany({
-    select: {
-      id: true,
-      body: true,
-      score: true,
-      member: {
-        select: {
-          id: true,
-          name: true,
-        },
-      },
-      store: {
-        select: {
-          id: true,
-          name: true,
-        },
-      },
-    },
-    where: {
-      storeId: store_id,
-      id: {
-        gt: cursor,
-      },
-    },
-    orderBy: {
-      id: "asc",
-    },
-    take: 5,
-  });
-
-  return reviews;
-};
-
 // 내가 작성한 리뷰 조회
 export const getMyReviews = async (memberId: number, cursor: number) => {
   const reviews = await prisma.review.findMany({
