+++ src/modules/reviews/services/review.service.ts
@@ -1,8 +1,8 @@
-import { IAddReviewRequest, responseFromReviews, ReviewListResponse, reviewResponse } from "../dtos/review.dto.js";
-import { addReview, getAllStoreReviews, getMyReviews, getReview } from "../repositories/review.repository.js";
+import { IAddReviewRequest, IReviewResponse } from "../dtos/review.dto.js";
+import { addReview, getMyReviews, getReview } from "../repositories/review.repository.js";
 
 // 리뷰 작성하기
-export const addStoreReview = async (data: IAddReviewRequest) => {
+export const addStoreReview = async (data: IAddReviewRequest): Promise<IReviewResponse> => {
   const reviewId = await addReview({
     memberId: data.memberId,
     storeId: data.storeId,
@@ -22,31 +22,16 @@ export const addStoreReview = async (data: IAddReviewRequest) => {
     throw new Error("리뷰를 찾을 수 없습니다.");
   }
 
-  return reviewResponse({
+  return {
     memberId: Number(review.memberId),
     storeId: Number(review.storeId),
-    body: review.body ?? "",
-    score: review.score ?? 0,
-  });
-};
-
-// 가게 리뷰 목록 조회
-export const listStoreReviews = async (storeId: number, cursor: number): Promise<ReviewListResponse> => {
-  const reviews = await getAllStoreReviews(storeId, cursor);
-
-  const mapped = reviews.map((r) => ({
-    id: Number(r.id),
-    body: r.body,
-    score: Number(r.score),
-    userName: r.member?.name || null,
-    storeName: r.store?.name || null,
-  }));
-
-  return responseFromReviews(mapped);
+    body: review.body || "",
+    score: review.score || 0,
+  };
 };
 
 // 내가 작성한 리뷰 조회
-export const handleMyReviews = async (memberId: number, cursor: number): Promise<ReviewListResponse> => {
+export const handleMyReviews = async (memberId: number, cursor: number) => {
   const reviews = await getMyReviews(memberId, cursor);
 
   const mapped = reviews.map((r) => ({
@@ -57,5 +42,10 @@ export const handleMyReviews = async (memberId: number, cursor: number): Promise
     storeName: r.store?.name || null,
   }));
 
-  return responseFromReviews(mapped);
+  return {
+    data: mapped,
+    pagination: {
+      cursor: mapped.at(-1)?.id ?? null,
+    },
+  };
 };
