+++ src/modules/stores/services/store.service.ts
@@ -1,5 +1,5 @@
-import { IAddStoreRequest, IStoreResponse } from "../dtos/store.dto.js";
-import { addStore, getStore } from "../repositories/store.repository.js";
+import { IAddStoreRequest, IStoreResponse, ReviewListResponse } from "../dtos/store.dto.js";
+import { addStore, getAllStoreReviews, getStore, getStoreMissions } from "../repositories/store.repository.js";
 
 export const addRegionStore = async (data: IAddStoreRequest): Promise<IStoreResponse> => {
   const storeId = await addStore({
@@ -22,5 +22,45 @@ export const addRegionStore = async (data: IAddStoreRequest): Promise<IStoreResp
     name: store.name || "",
     address: store.address || "",
     score: store.score ?? undefined,
+  };
+};
+
+// 가게 리뷰 목록 조회
+export const listStoreReviews = async (storeId: number, cursor: number): Promise<ReviewListResponse> => {
+  const reviews = await getAllStoreReviews(storeId, cursor);
+
+  const mapped = reviews.map((r) => ({
+    id: Number(r.id),
+    body: r.body,
+    score: Number(r.score),
+    userName: r.member?.name || null,
+    storeName: r.store?.name || null,
+  }));
+
+  return {
+    data: mapped,
+    pagination: {
+      cursor: mapped.at(-1)?.id ?? null,
+    },
+  };
+};
+
+// 특정 가게의 미션 목록 조회
+export const handleStoreMissions = async (storeId: number, cursor: number) => {
+  const missions = await getStoreMissions(storeId, cursor);
+
+  const mapped = missions.map((m) => ({
+    id: Number(m.id),
+    storeId: storeId,
+    reward: Number(m.reward),
+    deadline: m.deadline?.toISOString(),
+    missionSpec: m.mission_spec || "",
+  }));
+
+  return {
+    data: mapped,
+    pagination: {
+      cursor: mapped.at(-1)?.id ?? null,
+    },
   };
 };
